+++ src/views/index.ts
@@ -6,10 +6,6 @@
 export { PuckEditorView } from './PuckEditorView.js'
 export type { PuckEditorViewProps } from './PuckEditorView.js'
 
-// Client components
-export { PuckEditorClient } from './PuckEditorClient.js'
-export type { PuckEditorClientProps } from './PuckEditorClient.js'
-
 // Context
 export { PuckConfigProvider, usePuckConfig } from './PuckConfigContext.js'
 export type { PuckConfigProviderProps, PuckConfigContextValue } from './PuckConfigContext.js'